Remove individual image rotation
This can cause the images to not equal in size
which means further functions and script arguments
would be needed.
This script is intended more for 2 images taken using
side-by-side cameras, a beam splitter or a sliding rig.
If necessary, an image editor such as GIMP can be used
to do this, as well as equal sizing, color and lighting
correction, etc.